Cloud_pipelines/Azure/image-pipeline/requirements.txt
numpy
mxnet==1.1.0
opencv-python==3.1.0.5
azure-storage-blob==1.5.0
